Improve CMakeLists.txt
If the project is not the top level project, will append additional
flags to those set by the parent project.

If the project is the top level project, will append additional flags
to the default flags (which is saved during the first configure).

Setting of flags is no longer done on only the first configure. The
above improvements make it so that multiple configures will not keep
appending flags.  Running `make` after changing the flags in the
CMakeLists.txt will correctly generate new makefiles that use the new
flags then build the project using the new makefiles.