Shrink array capacity to half instead of a fourth
Still shrink them when their size drops to capacity/4 or lower. However,
shrink the capacity to capacity/2 instead of capacity/4. This prevents
the case where a pop causes a shrink and then a subsequent push causes
an expand. That is two resize operations in a row. The idea is to shrink
but still leave some space so that a subsequent push doesn't require an
expand.